requirements.txt
youtube-transcript-api
crewai
python-dotenv
streamlit